Share the expect-a-user-error helper between tests
Two tests had near-identical copies. Keep the one that checks the message
mentions what the test broke, so that a test can't pass on some unrelated error,
and give it an overload for callers that don't care what the message says.
